// src/app/models/user/save-profile-info.model.ts
export interface SaveProfileInfoModel {
  firstName?: string;
  lastName?: string;
  city?: string;
  email?: string;
  password?: string;
  phoneNumber?: string;
  values?: string;
  userId?: string;
  dateBirth?: string;
  patronymic?: string;
  inn?: string;
  pc?: string;
  kpp?: string;
  bik?: string;
  passportSerial?: number;
  passportNumber?: number;
  dateGive?: string;
  whoGive?: string;
  code?: string;
  addressRegister?: string;
  documentName?: string;
  countTimeSite?: string;
  countAd: number;
  defaultBankName?: string;
  corrAccountNumber?: string;
}